🎨 Fusion types in, ARCHE extension.
