+++ src/app/playground/core/character/Character.ts
