--- src/app/playground/core/extension/arche/types/CharacterState.ts
@@ -1,8 +1,0 @@
-// Used to check the character state.
-export interface CharacterState {
-  canJump: boolean;
-  inMotion: boolean;
-  slopeAngle: number | null;
-  characterRotated: boolean;
-  excludeRay?: boolean;
-}
