+++ src/app/playground/core/extension/arche/types/LockCamera.ts
