--- src/app/playground/core/extension/arche/types/customRigidBody.ts
@@ -1,7 +1,0 @@
-import { RapierRigidBody } from "@react-three/rapier";
-
-// RapierObject<RapierRigidBody> or <RapierRigidBody>
-export interface customRigidBody extends RapierRigidBody {
-  rotateCamera?: (x: number, y: number) => void;
-  rotateCharacterOnY?: (rad: number) => void;
-}
